\name{drawPCGPcircos}
\alias{drawPCGPcircos}
\title{
  Draw the circos plot for a single patient from the PCGP.
}
\description{
  This function will draw a circos plot for a single patient from the PCGP.  
}


\author{
  Steven W. Paugh \email{steven.paugh@stjude.org}
}


\examples{
drawPCGPcircos("SJTALL001")
}

\keyword{circos}

