\name{getPCGPgenedata}
\alias{getPCGPgenedata}
\title{
Retrieve gene level data from the PCGP data dashboard
}
\description{
This function will retrieve the gene level data from the PCGP data dashboard.  
}


\author{
Steven W. Paugh \email{steven.paugh@stjude.org}
}


\examples{
getPCGPgenedata("EZH2")
}

\keyword{sample list}

