\name{getPCGPgenelist}
\alias{getPCGPgenelist}
\title{
Retrieve the current gene symbol listing for the PCGP
}
\description{
This function will retrieve the a current gene symbol listing for the PCGP. 
}


\author{
Steven W. Paugh \email{steven.paugh@stjude.org}
}


\examples{
getPCGPgenelist()
}

\keyword{sample list}

