\name{getPCGPsamplelist}
\alias{getPCGPsamplelist}
\title{
  Retrieve the current patient listing for the PCGP
}
\description{
  This function will retrieve the anonymized patient identifiers and disease type for PCGP samples.  
}


\author{
  Steven W. Paugh \email{steven.paugh@stjude.org}
}


\examples{
getPCGPsamplelist()
}

\keyword{sample list}

