\name{PCGPData}
\alias{PCGPData}
\docType{package}
\title{
  PCGP Data R package
}
\description{
  The package serves as an API for PCGP data at explore.pediatriccancergenomeproject.org
}
\details{
  \tabular{ll}{
    Package: \tab PCGPData\cr
    Type: \tab Package\cr
    Version: \tab 0.4\cr
    Date: \tab 2012-12-31\cr
    Depends: R (>= 1.8.0), XML, png
    License: \tab \cr
    LazyLoad: \tab yes\cr
  }
  
}
\author{
  Steven W. Paugh <steven.paugh@stjude.org>
  Maintainer: Steven W. Paugh <steven.paugh@stjude.org>
}

\keyword{ package }
